Deleted ObjectPreviewSuspenseCacheAdapter and moved logic into bootstrapApp
