+++ src/ui/setup/index.ts
@@ -25,7 +25,10 @@ import { ThreadFront } from "protocol/thread";
 import { replayClient } from "shared/client/ReplayClientContext";
 import { ReplayClient } from "shared/client/ReplayClient";
 import { getPreferredLocation } from "ui/utils/preferredLocation";
-import { Location } from "@replayio/protocol";
+import { ExecutionPoint, Location, PauseData, PauseId } from "@replayio/protocol";
+import { addPauseDataListener } from "protocol/thread/pause";
+import { preCacheObjects } from "@bvaughn/src/suspense/ObjectPreviews";
+import { trackExecutionPointPauseIds } from "@bvaughn/src/suspense/PauseCache";
 
 declare global {
   interface Window {
@@ -106,6 +109,21 @@ export async function bootstrapApp() {
   if (typeof window === "undefined") {
     return store;
   }
+
+  // Connect data in legacy Redux stores (like PauseData or points) to the newer Suspense caches.
+  // This avoids requiring the new components from requesting redundant data.
+  // In the case of PauseData, it's extra important– because the backend won't re-send the same data twice no matter how many times you ask for it.
+  addPauseDataListener((pauseId: PauseId, executionPoint: ExecutionPoint, pauseData: PauseData) => {
+    const { objects } = pauseData;
+    if (objects) {
+      // Be sure to clone object data before pre-caching it.
+      // Otherwise ValueFronts might deeply mutate it and change its structure.
+      const cloned = JSON.parse(JSON.stringify(objects));
+      preCacheObjects(pauseId, cloned);
+    }
+
+    trackExecutionPointPauseIds(executionPoint, pauseId);
+  });
 
   // Wire up new Console and Object Inspector to the Redux logic for preferred source.
   (replayClient as ReplayClient).injectGetPreferredLocation(
